View transform redone in util
