--- app/transforms/fd-definition.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-flexberry-designer/transforms/fd-definition';
