+++ tests/unit/serializers/fd-dev-view-test.js
@@ -8,7 +8,6 @@ moduleForModel('fd-dev-view', 'Unit | Serializer | fd-dev-view', {
     'transform:decimal',
     'transform:guid',
 
-    'transform:fd-definition',
     'transform:i-c-s-soft-s-t-o-r-m-n-e-t-access-type',
     'transform:i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-t-write-mode',
     'transform:i-c-s-soft-s-t-o-r-m-n-e-t-business-data-service-object-events',
